requirements.txt
google-generativeai>=0.3.0
PyGithub>=2.1.1
